refactor: move types to api-kinetic-data-access
