+++ libs/api/account/data-access/src/index.ts
@@ -1,10 +1,6 @@
 export * from './lib/api-account-data-access.module'
 export * from './lib/api-account-data-access.service'
-export * from './lib/dto/close-account-request.dto'
 export * from './lib/dto/create-account-request.dto'
-export * from './lib/entities/account.info'
 export * from './lib/entities/balance-response.entity'
 export * from './lib/entities/balance-token.entity'
 export * from './lib/entities/history-response.entity'
-export * from './lib/entities/token-info.entity'
-export * from './lib/helpers/validate-close.account'
