+++ libs/api/kinetic/data-access/src/lib/api-kinetic.service.ts
@@ -2,17 +2,28 @@ import { ApiCoreDataAccessService, AppEnvironment } from '@kin-kinetic/api/core/
 import { parseTransactionError } from '@kin-kinetic/api/kinetic/util'
 import { ApiSolanaDataAccessService } from '@kin-kinetic/api/solana/data-access'
 import { ApiWebhookDataAccessService, WebhookType } from '@kin-kinetic/api/webhook/data-access'
-import { Commitment, removeDecimals, Solana } from '@kin-kinetic/solana'
+import { Keypair } from '@kin-kinetic/keypair'
+import {
+  Commitment,
+  generateCloseAccountTransaction,
+  getPublicKey,
+  PublicKeyString,
+  removeDecimals,
+  Solana,
+} from '@kin-kinetic/solana'
 import { BadRequestException, Injectable, Logger, OnModuleInit, UnauthorizedException } from '@nestjs/common'
 import { Counter } from '@opentelemetry/api-metrics'
 import { App, AppEnv, Prisma, Transaction, TransactionErrorType, TransactionStatus } from '@prisma/client'
 import { Transaction as SolanaTransaction } from '@solana/web3.js'
 import { Request } from 'express'
 import * as requestIp from 'request-ip'
+import { CloseAccountRequest } from './dto/close-account-request.dto'
+import { AccountInfo } from './entities/account.info'
 import { GetTransactionResponse } from './entities/get-transaction-response.entity'
 import { LatestBlockhashResponse } from './entities/latest-blockhash-response.entity'
 import { MinimumRentExemptionBalanceRequest } from './entities/minimum-rent-exemption-balance-request.dto'
 import { MinimumRentExemptionBalanceResponse } from './entities/minimum-rent-exemption-balance-response.entity'
+import { validateCloseAccount } from './helpers/validate-close.account'
 import { ProcessTransactionOptions } from './interfaces/process-transaction-options'
 import { TransactionWithErrors } from './interfaces/transaction-with-errors'
 
@@ -20,6 +31,9 @@ import { TransactionWithErrors } from './interfaces/transaction-with-errors'
 export class ApiKineticService implements OnModuleInit {
   private logger = new Logger(ApiKineticService.name)
 
+  private closeAccountRequestCounter: Counter
+  private closeAccountRequestInvalidCounter: Counter
+  private closeAccountRequestValidCounter: Counter
   private confirmSignatureFinalizedCounter: Counter
   private confirmTransactionSolanaConfirmedCounter: Counter
   private mintNotFoundErrorCounter: Counter
@@ -37,6 +51,15 @@ export class ApiKineticService implements OnModuleInit {
   ) {}
 
   onModuleInit() {
+    this.closeAccountRequestCounter = this.data.metrics.getCounter(`api_account_close_account_request`, {
+      description: 'Number of closeAccount requests',
+    })
+    this.closeAccountRequestInvalidCounter = this.data.metrics.getCounter(`api_account_close_account_invalid_request`, {
+      description: 'Number of invalid closeAccount requests',
+    })
+    this.closeAccountRequestValidCounter = this.data.metrics.getCounter(`api_account_close_account_valid_request`, {
+      description: 'Number of valid closeAccount requests',
+    })
     this.confirmSignatureFinalizedCounter = this.data.metrics.getCounter(
       `api_kinetic_confirm_signature_finalized_counter`,
       { description: 'Number of makeTransfer finalized Solana transactions' },
@@ -77,8 +100,133 @@ export class ApiKineticService implements OnModuleInit {
     return this.solana.deleteConnection(appKey)
   }
 
+  async getAccountInfo(
+    appKey: string,
+    accountId: PublicKeyString,
+    mint: PublicKeyString,
+    commitment: Commitment,
+  ): Promise<AccountInfo> {
+    const solana = await this.getSolanaConnection(appKey)
+    const account = getPublicKey(accountId)
+    const accountInfo = await solana.getParsedAccountInfo(account, commitment)
+
+    const parsed = accountInfo?.data?.parsed
+
+    const isMint = parsed?.type === 'mint'
+    const isTokenAccount = parsed?.type === 'account'
+
+    const owner = isTokenAccount ? parsed.info.owner : null
+
+    const result = {
+      account: account.toString(),
+      isMint,
+      isOwner: false,
+      isTokenAccount,
+      owner,
+      program: accountInfo?.owner?.toString() ?? null,
+      tokens: !isMint && !isTokenAccount ? [] : null,
+    }
+
+    // We only want to get the token accounts if the account is not a mint or token account
+    if (isMint || isTokenAccount) {
+      return result
+    }
+
+    const appEnv = await this.data.getAppEnvironmentByAppKey(appKey)
+    const appMint = this.validateMint(appEnv, appKey, mint.toString())
+
+    const tokenAccounts = await solana.getTokenAccounts(account, appMint.mint.address, commitment)
+
+    for (const tokenAccount of tokenAccounts) {
+      const info = await solana.getParsedAccountInfo(tokenAccount, commitment)
+      const parsed = info?.data?.parsed?.info
+
+      result.tokens.push({
+        account: tokenAccount,
+        balance: parsed?.tokenAmount?.amount ? removeDecimals(parsed.tokenAmount.amount, appMint.mint.decimals) : null,
+        closeAuthority: parsed?.closeAuthority ?? null,
+        decimals: appMint.mint.decimals ?? 0,
+        mint: appMint.mint.address,
+        owner: parsed?.owner ?? null,
+      })
+    }
+
+    return {
+      ...result,
+      isOwner: result.tokens.length > 0,
+    }
+  }
+
   getSolanaConnection(appKey: string): Promise<Solana> {
     return this.solana.getConnection(appKey)
+  }
+
+  async handleCloseAccount(
+    input: CloseAccountRequest,
+    {
+      appEnv,
+      appKey,
+      headers,
+      ip,
+      ua,
+    }: { appKey: string; appEnv: AppEnvironment; headers?: Record<string, string>; ip?: string; ua?: string },
+  ): Promise<Transaction> {
+    this.closeAccountRequestCounter.add(1, { appKey })
+    const accountInfo = await this.getAccountInfo(appKey, input.account, input.mint, input.commitment)
+
+    try {
+      const tokenAccount = validateCloseAccount({
+        info: accountInfo,
+        mint: input.mint,
+        mints: appEnv.mints.map((m) => m.mint?.address),
+        wallets: appEnv.wallets.map((w) => w.publicKey),
+      })
+
+      this.closeAccountRequestValidCounter.add(1, { appKey })
+
+      const mint = this.validateMint(appEnv, appKey, input.mint)
+
+      // Create the Transaction
+      const transaction: TransactionWithErrors = await this.createTransaction({
+        appEnvId: appEnv.id,
+        commitment: input.commitment,
+        ip,
+        referenceId: input.referenceId,
+        referenceType: input.referenceType,
+        ua,
+      })
+
+      const { blockhash, lastValidBlockHeight } = await this.getLatestBlockhash(appKey)
+
+      const signer = Keypair.fromSecret(mint.wallet?.secretKey)
+
+      const { transaction: solanaTransaction } = generateCloseAccountTransaction({
+        addMemo: mint.addMemo,
+        blockhash,
+        index: input.index,
+        lastValidBlockHeight,
+        signer: signer.solana,
+        tokenAccount: tokenAccount.account,
+      })
+
+      return this.processTransaction({
+        appEnv,
+        appKey,
+        blockhash,
+        commitment: input.commitment,
+        transaction,
+        decimals: mint?.mint?.decimals,
+        feePayer: tokenAccount.closeAuthority,
+        headers,
+        lastValidBlockHeight,
+        mintPublicKey: mint?.mint?.address,
+        solanaTransaction,
+        source: input.account,
+      })
+    } catch (error) {
+      this.closeAccountRequestInvalidCounter.add(1, { appKey })
+      throw error
+    }
   }
 
   createTransaction({
