+++ libs/api/kinetic/data-access/src/lib/entities/account.info.ts
