+++ libs/api/kinetic/data-access/src/lib/entities/token-info.entity.ts
