+++ libs/api/kinetic/data-access/src/lib/helpers/validate-close.account.ts
