+++ libs/api/queue/data-access/src/lib/queue/close-account/api-queue-close-account.processor.ts
@@ -1,5 +1,5 @@
-import { validateCloseAccount } from '@kin-kinetic/api/account/data-access'
 import { getAppKey } from '@kin-kinetic/api/core/util'
+import { ApiKineticService, validateCloseAccount } from '@kin-kinetic/api/kinetic/data-access'
 import { Commitment } from '@kin-kinetic/solana'
 import { Process, Processor } from '@nestjs/bull'
 import { Logger } from '@nestjs/common'
@@ -11,7 +11,7 @@ import { ApiQueueCloseAccountService } from './api-queue-close-account.service'
 export class ApiQueueCloseAccountProcessor {
   private readonly logger = new Logger(ApiQueueCloseAccountProcessor.name)
 
-  constructor(private readonly service: ApiQueueCloseAccountService) {}
+  constructor(private readonly kinetic: ApiKineticService, private readonly service: ApiQueueCloseAccountService) {}
 
   @Process(QueueOptions[QueueType.CloseAccount])
   async handleProcess(job: Job, cb: DoneCallback) {
@@ -22,7 +22,7 @@ export class ApiQueueCloseAccountProcessor {
 
     try {
       const appKey = getAppKey(environment, index)
-      const accountInfo = await this.service.account.getAccountInfo(appKey, account, mint, commitment)
+      const accountInfo = await this.kinetic.getAccountInfo(appKey, account, mint, commitment)
 
       try {
         validateCloseAccount({ info: accountInfo, mint, mints, wallets })
@@ -32,7 +32,7 @@ export class ApiQueueCloseAccountProcessor {
         )
 
         const appEnv = await this.service.data.getAppEnvironmentByAppKey(appKey)
-        const transaction = await this.service.account.handleCloseAccount(
+        const transaction = await this.kinetic.handleCloseAccount(
           {
             commitment,
             mint,
